// frontend/src/backend.d.ts
export interface BusinessProfile {
    offerType: OfferType;
    seasonalActivity: SeasonalActivity;
    numberOfSeats: bigint;
    location: string;
}
// ...
export enum OfferType {
    bar = "bar",
    brzaHrana = "brzaHrana",
    kafic = "kafic",
    ostalo = "ostalo",
    restoran = "restoran"
}
export enum PaymentMethod {
    kartica = "kartica",
    gotovina = "gotovina"
}
export enum SeasonalActivity {
    oboje = "oboje",
    zima = "zima",
    ljeto = "ljeto"
}